- sweep for warnings

+++ copasi/MIRIAMUI/CQBiologicalDescriptionDM.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2021 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -54,7 +54,7 @@ size_t  CQBiologicalDescriptionDM::size() const
 
 int CQBiologicalDescriptionDM::rowCount(const QModelIndex& C_UNUSED(parent)) const
 {
-  return mFetched + 1;
+  return (int)mFetched + 1;
 }
 
 int CQBiologicalDescriptionDM::columnCount(const QModelIndex& C_UNUSED(parent)) const
@@ -208,7 +208,7 @@ bool CQBiologicalDescriptionDM::removeRows(int position, int rows, const QModelI
   if (rows <= 0)
     return true;
 
-  beginRemoveRows(parent, position, std::min< int >(mFetched, position + rows) - 1);
+  beginRemoveRows(parent, position, std::min< int >((int)mFetched, position + rows) - 1);
 
   std::vector< const CBiologicalDescription * > ToBeDeleted;
   ToBeDeleted.resize(rows);
