+++ copasi/MIRIAMUI/CQReferenceDM.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2021 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -55,7 +55,7 @@ size_t CQReferenceDM::size() const
 
 int CQReferenceDM::rowCount(const QModelIndex& C_UNUSED(parent)) const
 {
-  return mFetched + 1;
+  return (int)mFetched + 1;
 }
 
 int CQReferenceDM::columnCount(const QModelIndex& C_UNUSED(parent)) const
@@ -209,7 +209,7 @@ bool CQReferenceDM::removeRows(int position, int rows, const QModelIndex & paren
   if (rows <= 0)
     return true;
 
-  beginRemoveRows(parent, position, std::min< int >(mFetched, position + rows) - 1);
+  beginRemoveRows(parent, position, std::min< int >((int)mFetched, position + rows) - 1);
 
   std::vector< const CReference * > ToBeDeleted;
   ToBeDeleted.resize(rows);
