+++ copasi/barChart/CQ3DBarsModifier.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2021 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -7,10 +7,6 @@
 // Properties, Inc., University of Heidelberg, and University of
 // of Connecticut School of Medicine.
 // All rights reserved.
-
-
-
-
 
 #include  "CQ3DBarsModifier.h"
 
@@ -134,7 +130,6 @@ CQ3DBarsModifier::CQ3DBarsModifier(CQArrayAnnotationsWidget* widget, Q3DBars *ba
   m_graph->setActiveInputHandler(m_inputHandler);
   connect(m_inputHandler, SIGNAL(signalShowContextMenu(const QPoint &)), widget, SLOT(slotShowContextMenu(const QPoint&)));
   connect(m_inputHandler, SIGNAL(signalBarDoubleClicked(int, int)), widget, SLOT(selectTableCell(int, int)));
-
 }
 
 CQ3DBarsModifier::~CQ3DBarsModifier()
@@ -274,7 +269,7 @@ void CQ3DBarsModifier::actionTriggered(QAction* action)
 
   if (title == "Theme")
     {
-      changeTheme(menu->actions().indexOf(action));
+      changeTheme((int)menu->actions().indexOf(action));
     }
   else if (title == "Selection Mode")
     {
@@ -353,7 +348,6 @@ void CQ3DBarsModifier::actionTriggered(QAction* action)
         {
           m_primarySeries->setMesh(QAbstract3DSeries::MeshSphere);
         }
-
     }
   else if (title == "Shadow")
     {
@@ -386,7 +380,6 @@ void CQ3DBarsModifier::actionTriggered(QAction* action)
           m_graph->setShadowQuality(QAbstract3DGraph::ShadowQualitySoftHigh);
         }
     }
-
 }
 
 void CQ3DBarsModifier::changeRange(int range)
@@ -424,7 +417,6 @@ void CQ3DBarsModifier::changePresetCamera()
 
   if (++preset > Q3DCamera::CameraPresetDirectlyBelow)
     preset = Q3DCamera::CameraPresetFrontLow;
-
 }
 
 void CQ3DBarsModifier::changeTheme(int theme)
@@ -633,8 +625,6 @@ void CQ3DBarsModifier::setReflection(bool enabled)
   m_graph->setReflection(enabled);
 }
 
-
-
 CQCustomInputHandler::CQCustomInputHandler(DATAVIS_NS_PREFIX QAbstract3DGraph * graph, QObject * parent)
   : DATAVIS_NS_PREFIX QTouch3DInputHandler(parent)
   , m_graph(graph)
@@ -643,7 +633,6 @@ CQCustomInputHandler::CQCustomInputHandler(DATAVIS_NS_PREFIX QAbstract3DGraph * 
   , m_axisZ(NULL)
   , mState(CQCustomInputHandler::StateNormal)
 {
-
 }
 
 void CQCustomInputHandler::mouseDoubleClickEvent(QMouseEvent * event)
